First Commit
Opening a new repo for Scavenge and Survive, mostly because having ALL
my SA:MP gamemodes in the same folder kind of annoys me! All external
dependencies are stored in a Pawno folder that isn't on this repo, all
of those scripts and plugins are available from their respectable
authors and links are available in my main gamemode script.

Anything else that's required is located inside the /scripts/ directory,
if it's missing anything, let me know as soon as possible!