Changed some item behaviour and data
Changed "Patrol Pack" item name to "Small Bag", changed box names to
better suit their in-game models.

Changed the way players interact with bags, tapping F on a bag on the
floor now enters it's inventory and holding F will pick it up, similar
to how Safeboxes work.

Entering the inventory of safeboxes and bags will now play a crouch-in
animation and when the player exits the menu a crouch-to-stand animation
will play.

Added a test filterscript for random test code.