Fixed the get command to be more restrictive
I really must clean up that PlayerFunctions.pwn file, it's full of crap!

In fact I want to completely re-do the whole naming scheme of the
utilities functions such as math etc. I might just use an external
library if there is one that fits my needs.

The actual used player functions are just messages, distance checks, and
stuff like that, most of the stuff in that file is just a collection of
things I thought *might* be useful in the future when I used to develop
the freeroam script.