Added melee items removed some defines and prints
Added melee definitions for items, along with a melee animation-set
script.

Removed some old definitions and macros that weren't being used.

Added a backup feature for if the server crashes. Prior to this fix, if
the server crashed while saving safeboxes then all safeboxes would be
saved with item ID 0 (orignally a dummy flare item) would be saved
instead. Now if item ID 0 is saved to safeboxes, inventories or vehicles
the entire process is cancelled. If ID 0 is loaded from any of these it
won't be created. If an item with type ID 0 is created it will be
cancelled by hook code in scripts/SSS/DisallowActions.pwn.