Updated SIF functions, added bodies and stats
A torso item will be created at a player's death point, when this item
is interacted with a dialog will display how the player died. The torso
can be picked up by holding the interact key (needs a tooltip)

Updated all code for the latest SIF update that changes the names and
parameter orders of the container-item interaction callbacks (it's now
non-player dependent)

The above update allows inventories and containers to display weapon
ammunition as a bracketed number next to the item name (Similar to how
petrol cans display the amount of petrol inside)

Added a new script that will be later used for data collection, two
statistics are currently in place: UnarmedKills and ArmedKills. This
will be used to measure the most and least used items and weapons which
will later affect spawn rates.

Knock-out code has been moved to a separate function and script,
knockouts can now be induced by calling the function and "waking up" the
player can also be triggered via a function.