Added new bag type, improved parachute item
The parachute now acts as an item while held and can be equipped by
pressing Y (similar to bags) and can only be equipped when the player
has no bag.

Fixed small animation bug with the Phone Bomb item.

Completed the DestroyDefense function in Defenses to properly remove the
file.

Updated Weapons script to support the parachute improvement.

Added Parachute Bag item which can be created by combining a knife with
a parachute, with the new Craft update in SIF the knife will be returned
and the original parachute deleted. This new bag has 6 slots.