Added report functionality
Players can now be reported, unread reports will be sent to admins when
they log in. Reports can be managed by deleting or marking as unread, as
well as banning the player who was reported.

Moved "/setadmin" and "/setvip" to level 3.

Added "/aliases" command to check a player's other account names.

Made "/spec" level 2 and improved function.

Made "/goto" level 3 so it can't be abused.

Added loot spawns to Flint County.

Improved MovementDetection.pwn code, it still needs some work and
probably a distance threshold increase for when the server to player
connection is slow (This can result in incorrect accusation of hacking)

Added filterscript "objtoloot.pwn" that will convert a set of
CreateObject lines from a map editor into loot spawns. The loot index
type depends on the model ID (List of model IDs for each index in the
definitions at the top)