Week update, new items, combinations and features
Renamed "Beer" item to "Bottle", this expands the possibilities for
uses.

Added a "Watch" feature, this displays the time, bearing and radio
frequency. It also covers up the mini-map more effectively than the last
version, this also makes it impossible to see car icons on the mini-map.

Moved player time updating to a per-player timer rather than the global
one, mainly because of the new per-player watch UI.

Added a "Hands Up" animation, accessible by pressing Sprint and Crouch
together (Default is Space and C, which makes the animation easy to use
in tight situations)

Maps script now only loads .map files, any other extension will be
ignored.

Added an animation to putting items into a bag, this fixes the bug that
would make a player un-holster their weapon accidentally when putting
items in their bag by pressing Y.

Fixed an item name bug with the Clothes item.

Hidden the "/admins" command from players so finding hackers is easier,
hackers won't know if an admin is online or not so they won't be able to
stop hacking whenever admins are online.

Limited the duty ability by adding a 10 second cooldown after use.

Fixed an item duplication bug with duty mode.

Added the "Hands Up" animation to the global disallow actions code.

Added a debug filterscript "animtimer" which displays the time an
animation plays for as well as the animation name, library and index
number.

Added "Disarm.pwn" script that allows players to remove the weapons of a
player who is knocked out or surrendering.