Added more saved data to characters
New information about player characters is saved.

Stance, crouching, sitting or surrendering is saved.
Bleeding, you can't re-log to avoid bleeding, it is now saved.
Handcuffs, if you are handcuffed, it will be saved.

Fixed a bug where a character's weapon wouldn't save if it was saved
while that character was in a vehicle.

Being in a vehicle now only disables your weapon if you have a
"drive-by" weapon.