Two new features, bug fixes and code cleanup
Changed GeoIP plugin to an include version by Whitetiger because the
plugin version links are dead and this version may be better.
(IMPORTANT: Rename global variable 'result' inside GeoIP include by
Whitetiger. Absolutely terrible name for a global variable I must say.)

Added 'fixes2' by Y_Less to help fix a sa-mp console error that prevents
players from connecting.

Added ping limit global variable, value automatically updates when the
player count goes above or below 10. (600ms for below 10 450ms for 10 or
above) Admins can set the ping limit with the command "/setpinglimit".

Changed weather data array, it was stupid. (Feel free to mock the
original, it had cells filled with values that were just values of those
cell addresses)

Added item Tent Pack, when a hammer is used with this item while it's in
the world, a tent will be created there. Tents save on server
close/restart just like signs, and are not included in auto-saves. Tents
currently have no real function apart from camouflage for boxes or just
decoration for a camp.

Added item Campfire, an odd item that acts slightly different to other
items. It's item type model is blank, when picked up by a player it will
create a resized version of the log pile in their hands. When dropped,
the item will destroy and a campfire will be created in it's place. When
a player presses F at a campfire, it will be deleted and an item created
in it's place and given to the player (with the correct animation and
timing)
Campfire objects can be fuelled with a petrol can and lit with a
lighter. A campfire can be lit without fuel but has a 99.5% chance of
failing. Campfires have a 40% chance of lighting if the weather is rain.

Added experimental script "GlitchRestart" which is an attempt to fix the
"Kicked <IP> because they didn't logon to the game" error that sometimes
means the server has stalled. However I don't know if it works or not as
the error is sporadic.

Moved vehicle interaction code to "/scripts/SSS/Vehicle/Core.pwn".

Added correct hold offsets for the ammo tin item.

Added anti "g-bug" script that disables entering vehicles as a passenger
if there is no driver. This prevents players from using the G key to
enter a vehicle in an otherwise inaccessible area (possibly a player
built fort) to gain access to that area.

Decreased the amount of ammo spawned inside ammo tins to one magazine of
the weapon it's compatible with.

Fixed a but where the armour item would break player armour value if
used with full armour or after armour has been used (player has negative
armour)

Added CancelPlayerMovement to sign editing and placing to prevent
vehicle entry.

Added loot spawns to Red County (Only the cities LS and LV remain for
loot)

Removed admin's name from kick and ban messages to conceal the admin's
identity if they are undercover.

Added all new items (except campfire) to loot spawn index.

Rewritten the vehicle grouping code, all vehicles are now added to
groups and the large arrays are gone. Groups are now identified by a
single value in the vehicle data array.

Vehicles now have individual spawn rates, most are 1.0 but a lot of rare
and military vehicles have been lowered.

Fixed some major bugs with vehicle saving.

Fixed bugs with weapon reloading and equipping weapons from inside
containers.

Fixed some major bugs with the autosave code in each of the main
autosaved entites (defences, vehicles and safeboxes)

Added a "modded" data file for loot, this data file replaces all loot
spawn indexes with 100% weapons, meds and armour for all out chaos.