Ammo and Admin feature update
Added new ammo tin items, there is now an item for each ammo type
instead of one item that contains the type and amount in it's extra data
(which involved some awkward bitwise trickery) this way, all the extra
data has to store is the amount of ammo in the tin and each ammo type
has a unique model so it's easier for players to spot a specific type of
ammo.

Added a spawn tick variable for checking when the player spawned (may be
used in future statistics features, "days alive" etc)

Moved ban code to new file "/scripts/SSS/Cmds/Ban.pwn" this now contains
a ShowListOfBans function used in a moderator level command.

Fixed the Molotov Cocktail item combination creation code, there was a
bug due to the new weapon code.

Added ability to move ammunition between guns and tins and all
combinations (gun to gun, tin to tin, gun to tin, tin to gun)

Fixed a small bug with moving a holstered weapon to inventory.

Fixed the /acmds command lists to display the right (and existing)
commands.

Added "/up (distance)" and "/ford (distance)" for on-duty admins.

Added return value of 6 to all duty commands if the player isn't on
duty. This displays a message on OnPlayerCommandText in the master
script.

Removed weather from moderator level and put it to developer level, it's
not really required and is never used, if moderators could change the
weather whenever they please, they'd probably only do it for their own
benefit anyway.

Temporarily removed ammo tins from loot spawns until the code is
finished.

Fixed a bug where if a player died then quit or experienced a server
restart, they would log back in alive.

Removed fort parts from autosaves, their state changes can be tracked
easily anyway. Fort parts are saved as soon as they are constructed or
as soon as a player modifies them, and removed when they are destroyed.

Optimised autosaves some more, players still take ~100ms to save each
now, thanks to SQLite (maybe removed in a future update)

Improved the hack detection somewhat and fixed some bugs that were
sending false reports.