Ammo, inventory, loot and fort fixes
Fixed some small bugs and exploits in ammo tin usage.

Fixed a major bug that resulted in holstered weapon loss if moved to a
full inventory.

Lowered spawn rate for bandages and medkits, raised spawn rate for
injectors and pills so they are used more often, plan to remove
unlabeled pills and empty injectors from spawns.

Fixed a bug where GetAmmunitionRemainder would return negative values
which is incorrect.

Fixed a bug where defence fort part keycode wouldn't save as soon as it
is entered, if the server crashed or was closed incorrectly the keycode
would be "0000".