BBQ, Campfire and Food bug fixes
Moved OnItemNameRender hook for food items to Food.pwn (it was in the
BBQ script)

Improved Action Text messages for campfire to show more useful
information to the user.

Disabled adding of already cooked food to BBQs.

Fixed a bug where putting a second food item on a BBQ would give the
first back (swap the food items)

Fixed some bugs with picking up a BBQ while it had food items on.

Improved campfire functionality a lot, added a smoke particle effect
that appears when the food is cooked.

Fixed some bugs where the food could be cooked even after being removed
from the campfire or when the campfire has already burnt out.