New items and item updates
Added a fun new item that takes advantage of a seemingly broken
animation in SA.

Added range limits to the entity deletion commands.

Added items to loot spawn indexes that are required to combine with
other items. Some were removed a while ago due to being useless.

Added a receiver item at the Mt. Chilliad radio station which is the
only item of it's type in the entire map and is requires to construct a
Hack Device.