Fixed an item duplication bug with tents
Fixed a bug where placing an item in range of two tents would save the
item twice resulting in duplicating the item after a restart. It now
deletes the item once it is saved, because this happens at the restart
and there is no autosave for tents, deleting the item is no issue.