Inventory-weapon bug fixes
Fixed a bug where if a player moved a weapon item from their inventory
to their bag it would be equipped instead, and if the player was already
holding something it would be deleted or they would be given the item
and the weapon at the same time.

Fixed a bug where weapon items in the players inventory would not load
correctly as the holster script was disallowing holsterable weapons to
be put in the inventory.

Fixed a minor problem with tents saving items they shouldn't (sometimes
items get stuck in the centre of the map).

The second issue was resolved with an update to the SIF source, check
that repository for the latest version and more details on this update.