Updated vehicles, include order, new hack detect
Re-written a lot of the vehicle spawn code, vehicle spawning is now
easier to manage and change in the code, it also stores more data about
vehicles on the server which can be used in the future to catch hacks
such as tire fixing.

Moved "/Data/" and some "/Core/" scripts above the "/Player/" folder and
used interface functions for player data instead of direct variable
addressing.

Moved "MAX_SPAWNED_VEHICLES" to the global scope and increased it to
150.

Renamed all vehicle variables and made them private/local to only the
vehicle script with interface functions for manipulating.

Added a new hack detection feature that detects players who teleport
vehicles to them. This feature is still in testing.

Moved all player vehicle related code to a new file.

Removed code that destroys the items in a container before the container
is destroyed, the container destroy function does this anyway and this
might have caused some cases of incorrect items being destroyed.