Anti vehicle surf, vehicle bug fixes
Added a new function IsValidVehicleID to replace IsValidVehicle which
performs an additional check to ensure the ID is below the server limit
(not the SA:MP limit)

Reorganised damage code to ensure bleeding is applied correctly while
using server-side custom weapons.

Added an extra check to vehicle teleporting to ensure a player is only
reported if the teleported vehicle is within a certain distance of the
player.

Added "VTP" and "REPORT_TYPE_CARTELE" to the reports script.

Added "VehicleSurfingCheck" update function, this feature will throw the
player off, knock them out and remove some health (randomly 0 to 5) if
the vehicle they are standing on goes over 40Km/h

Fixed some bugs with vehicle owners not being reset. Now the new
function SetVehicleOwner will automatically remove owner names from
vehicles if a new player is assigned to that vehicle or if the owner
saves another vehicle.

Added constant "VEHICLE_SPAWN_CHANCE" to easily change the chance of a
vehicle spawning (in percent) default is 4%.

Fixed a bug in defences that wouldn't remove defence part hitpoints
properly.

Fixed a bug with repair tools where the player could put the tool in
their inventory or remove it some other way and the vehicle would keep
getting fixed all the way to 1000.0 hp.

Removed some unwanted includes, rewritten the comments on some that are
now on Gist. Will probably remove all filterscripts from the repo and
just upload the maps one (or integrate it to the main code) then upload
the released filterscripts to Gist.