Campfire bugfix, weapon magazine limit update
Removed a pointless check in Barbecue.

Fixed a bug where the campfire wouldn't actually use or check for any
fuel in the fuel can.

Added an extra check for if the food item is valid when cooking has
finished.

Fixed a bug where the cooking timer would keep going even if the food
was taken off the fire before it cooked.

Fixed a bug where the fire would become unusable after taking off the
first food item by changing the pickup callback from
OnPlayerPickedUpItem to OnPlayerPickUpItem, this was making the SIF item
script miss setting the "interactor" variable for the item back to
INVALID_PLAYER_ID.

Changed the maximum amount of magazines for weapons to better reflect a
realistic and challenging limit to move combat and gameplay away from
weapons by making the ammo more scarce and precious.