Fixed a possible bug with safeboxes
(Not sure if fixed, don't know how to reproduce at all, but recoded this
anyway in hope it would fix the issue. The issue was randomly
disappearing boxes.)

Removed "file_OS" call from OnGameModeInit, the newest version of
"IniFiles" (by me) hooks OnGameModeInit automatically for this feature
(OS detection).

Removed "DestroySafebox" and moved the code straight into OnItemDestroy.
This should also fix a possible problem with boxes randomly
disappearing. I think what was happening was, a box in the world was
destroyed, then DestroyItem was called AGAIN inside "DestroySafebox" (no
idea why) which triggered a recursion that deleted a lot of other boxes.

Destroying the item is all that you need to do to remove a box anyway,
no idea why I wrote a special function specifically for deleting a
certain item type.