Added new puzzle type, added save-blocking
Added a new item for the new puzzle, a "code part" displays a single
digit value as it's extra-data value, these can be found to complete a
keycode used to enter an area.

Added a branding UI element with my blog URL on.

Updated OnPlayerKeypadEnter to be hooked in the Area69 script (it wasn't
hooked before because it was only used there)

Added "SaveBlock" script, this manipulates input coordinates to a
"reset" position if the coordinates are inside any areas defined for
"save-blocking". This is to prevent players from staying inside certain
areas to farm loot or stay safe from others.

Added "Code Hunt" puzzle, in each zone a "lockup" will be available to
access via entering a 4 digit code. The digits for this code are
scattered around the area in random places. The first Code Hunt is set
in San Fierro and features 10 possible locations for code parts.

Renamed "main.pwn" to "testing.pwn" latest test is a function to pick
some random numbers from a finite set (this function can be found inside
"utils\math.pwn")