Added armour model, rcon whitelist on/off control
Added "on" and "off" parameters to RCON "whitelist" command.

Added an armour model that attaches to the player when they have some
armour.

Increased the effectiveness of armour.

Added a small fix for a camera detection bug in "HackDetect.pwn"

Fixed a bug in "/report" command menu under "Specific player" list that
would show the reporters name not a list of online players.

Fixed a small bug with settings that would invert the effect for
"gPauseMap".

Fixed a bug with bicycles engines being off (GTA logic!)