Added tooltips for items, fixed some bugs
Note: Using the latest version of SIF, compilation with an older version
will result in errors due to a few small name changes.

Added a tooltip for every item type, this is a short description of what
the item does that is displayed to the player (if they have tooltips
enabled) when they pick up an item.

Added command "/tooltips" to toggle tooltip display.

Added extra checks to character data saving to ensure characters aren't
saved while in the connection screen overlooking Verona Beach (while the
character is placed at Verdant Bluffs)

Added an animation to the dropping of dice items and improved the
randomness of them with a set of rotations for each side instead of just
randomised rotations (that resulted in a higher chance to land on
certain sides)

Added a small debug hook for DestroyVehicle to pinpoint where in the
script a vehicle is destroyed from, in an attempt to find out of random
vehicle disappearance is a bug with the script or SA:MP itself.