Fixed a bug with vehicles driving without fuel
Restructured a lot of code inside 'PlayerVehicleUpdate' including some
UI elements, such as not displaying "DOR" when on a bike, and not
displaying fuel levels while in a vehicle that doesn't run on fuel.

Also fixed a bug that was very rare but allowed vehicles to be driven
without fuel. Now, the fuel check and engine off function are called
outside of the engine-on check meaning the engine off feature will
repeat while the player is in a vehicle without fuel.