Explosive items renamed and all coded
Renamed the "Trip Mine" to the "Proximity Mine" and renamed the "Motion
Mine" to the "Trip Mine", due to the change in the use of the Motion
Sensor and Accelerometer item to craft these. This change was to avoid
confusion with the name change.

Added a "combat-log-window" settings.cfg line which is the time in
seconds after a player gets shot that it is counted as a combat log if
they quit.

Added relational indexes for: Vehicles, Safeboxes and Bags to easily
acquire the container IDs of these entities. This is used in the
explosives update as an alternative to looping through every entity to
check if the explosive item is inside its container.

Updated the code for all explosives to use the new "SetItemToExplode"
function, which takes into account item storage. If an item is inside a
vehicle trunk and explodes, it will create the explosion at that vehicle
and destroy the vehicle too, same goes for boxes, bags, inventories and
world containers. All this code is inside
"scripts/Core/World/Explosive.pwn".

Added the proximity mine explosive type for EMP, IED and TNT explosives.
This type explodes when someone walks within 5m of it.