Converted all SQL database queries to SQLitei
Total database revamp:

Updated field name definitions so that there is a unique set per
database.
Added a new database for world objects, may in future move
vehicles/boxes/tents/etc here.
Updated database fields to have types now.

Moved Spraytags to "/World/", they are now more integrated into SS so
this makes sense. Also updated them to utilise the "World.db" database
to save spraytag data.

Disabled the handsup animation if the player isn't idle, there was an
exploit involving using the handsup animation to reset velocity to allow
for long falls.

Moved login code to a separate function which returns 3 different values
to determine the login result: failed, success, success and autologin.

Improved whitelist commands, added a "?" parameter that will check if a
name is in the whitelist. Also updated add/remove functions to return 0
if the name is already in/not in the whitelist respectively.

Fixed bugs with the shield protecting from the wrong side and not
protecting at all when holstered.

Added "gLoginFreezeTime" as a server setting in
"/scriptfiles/SSS/settings.cfg" which determines how long the player is
frozen for after logging in.

Added a utility function for converting a datetime unix timestamp into a
string of format "Day Date Month Year at Hour:Minute:Second".