Reorganised player database + variables
Added a "website" setting for whitelist message.

Changes to "Player" database table:
Removed "gender" column
Added: regdate, lastlog, spawntime, spawns, warnings.
spawntime is the unix timestamp of when the character was created (not
the account)

Added a "Profile" script that allows players to view other players
information such as IP, spawns, days survived etc.

Moved all item scripts from "/scripts/Items" to "/scripts/Core/Item/".

Renamed the world script to simply just "World" this means editing the
world just means editing that script file not the master script.

Added some checks to ensure gMapName and GenerateSpawnPoint are defined.

Added a line that prints when the server has finished saving.

Reorganised E_PLAYER_DATA structure, renamed "bPlayerGameSettings" to
simply "gPlayerBitData" and added a tag to it.

Added a "WhitelistKick" function that displays the whitelist
notification and kicks the player. Formats in the "gWebsiteURL"
variable.

Rebuilt the "ResetVariables" function for players.

Added spawn position to the player file, will be removed from the
database soon.

Fixed a lot of bugs introduced in the previous commit.