Ban bugfix, tent item bugfix, +more small fixes
Fixed a bug in the ban info not displaying when a banned player
connects.

Fixed a bug with clothes displaying the wrong gender.

Fixed a bug with the inventory script displaying weapons the player
isn't holding (for instance, parachutes after use)

Added an "IsItemLoot" function as well as "loot_IsItemLoot" variable for
every item. When items are created in loot spawns, this variable is
true. This helps in determining whether randomised extradata should be
applied to an item.

Added the "IsItemLoot" check to all items that are given randomised
extradata when they are created.

Added a feedback message for interacting with torsos with invalid IDs or
out of the harvest timeframe.

Added a function "IsItemTypeExtraDataDependent" which returns true if
the specified itemtype relies on it's extradata to function, such as
boxes and bags which use this value to address the containerid they are
assigned to.

Fixed extradata values being applied to items loaded from tents.