Moved all scripts, renamed master script
Shifted scripts to a new directory in the "gamemodes" folder, this helps
fix a bunch of issues with the compiler not reading files. These issues
became apparent in a recent update that moved all item scripts to the
"/Core/" directory.

Updated all include lines to reflect this change.

Also renamed the master script to "ScavengeSurvive" as "sss" wasn't very
descriptive.

Removed "API" scripts, these libraries aren't actually integrated into
SS in any way, so they will be released as separate libraries for use in
the "/pawno/include/" directory. (Links will be put in soon)