Fixed SA:MP vanishing vehicles +equip bug
Added a workaround for a SA:MP issue that caused vehicles to randomly
delete if parked on oil rigs or islands outside the map boundaries. The
workaround just recreates the vehicle and copies all the previous
vehicle data such as health and damage.

Fixed a bug where newly registering players wouldn't be inserted into
the database.

Fixed a bug where weapons would vanish when equipped from containers.

Removed the old tooltip from defence parts, now that the "/tooltip"
command is added with tooltips for all items.

Added a "/vreset" dev command that resets a vehicle (*doesn't respawn
it) this has the same effect as the vanishing vehicles fix. It just
destroys the vehicle and creates a new one with the same data at the
same position.