Fixed a bug with loot not spawning with extradata
(Requires a SIF update)

Fixed a bug with loot items not spawning with any extra-data that is
usually set within OnItemCreated. Since the fix for tents, inside this
callback is an "IsItemLoot" check that ensures the item is a loot spawn
before applying randomised data. This "IsItemLoot" property was being
applied to the item *after* it was created. Now, I've added a new
function to SIF/Item to predict the next itemid that will be used in
order to flag the item as loot before it is created and thus before
OnItemCreated is called so it can apply extra-data correctly.