Added easier print control on load/save functions
Load and Save functions for Vehicles, Safeboxes, Tents and Defences now
have two parameters: printeach and printtotal.
printeach, if true, will print to the console each time an entity of
that type is saved/loaded and printtotal will print the total number of
entities of which type were saved/loaded.

You may not want to have these on all the time, just for debugging.
However, the printtotal parameters only use one line so they aren't as
log-consuming as the printeach lines.