Added supply drops, tweaked military spawn rates
Added supply drops that fall on random locations once every hour or so
(interval is random from 1 hour to 1 hour 30 minutes) Drops consist of a
crate and a parachute and when they drop, a container is created with
some random military loot inside. I plan to make a separate loot index
just for supply drops, that contain medical supplies and cooked food as
well as some weapons.

Added supply drop locations all around the map, 38 in total. May add
more in future, with multiple drop positions per zone.

Tweaked military loot spawn rates slightly, weapons are a bit more
common and shields are a bit more rare. This change is to stop a lot of
the same item spawning in areas (such as helmets and shields)