Added auto-create directories and files
If any folders are missing (or the settings.cfg file) the server will
create them (and fill the settings.cfg with default values)

Changed all "fremove" to "file_remove", as it seems to be a lot more
reliable.

Changed sign load/save code to match that of other world entities.