Changed inventory layout, updated armour code
Added a "Face" tile to the inventory screen and changed the layout. This
currently has no function but it will be used in future when facial
items such as masks are added.

Added the ability to remove armour by clicking the "Torso" inventory
tile while viewing a container.

Made armour variable interface-function addressed as oppose to directly
addressed, changing the value now causes the armour item to disappear if
set to 0 and avoid allowing the value going out of bounds (<0 or >100).

Added a name-extra to armour items that displays the condition.

Velocity is now updated while on foot too.