Updated configuration to use DJson
Switched from INI format to JSON format for the server configuration
file. The JSON format is much more advanced and offers more features
such as arrays.

Settings are now organised into two classes, 'server' and 'player'.

Added a new configuration key, "infomsgs" which is an array of strings.
These strings are sent to players every 10 minutes and it cycles through
the messages. This means the link to Empire Bay and the wiki site are
finally removed from the actual source code!

Also fixed a typo with a vehicle spawn debug print.