Added auto-bans, fixed admin chat logging issue
Added auto-bans for various report types. The events are still reported
as reports contain a lot more information such as locations and report
type-specific data.
The report types that auto-ban are: Swim-fly, Vehicle health, Nitro and
Hydraulics.

Fixed a bug with admin chat logging that resulted in duplicate logging
(the log call was mistakenly inside a loop)