Lots of small fixes
Fixed a bug with gender not being assigned to new female characters.

Fixed a bug with the knockout bar not filling up while in a vehicle.

Fixed a bug with disarming a player while near another player, it
disarms them both but deletes one of the retrieved items.

Fixed an issue with spectating, made the spectate mode exit if the
target player leaves the server.

Updated the "animtimer" utility filterscript.

Fixed a bug with the InfoMessage timer, changed it from a task to just a
repeating timer because y_timers initiated before the interval was
loaded from the settings.json file.