Added car locks, fixed small issues
Added a "key" value for vehicles. This can be a positive integer between
0 and 2^31.

Improved vehicle interactions by adding callback
OnPlayerInteractVehicle. This is called when they press F/Enter near a
vehicle. The looping system for finding which vehicle the player is near
will be optimised in the future.

Moved crowbar-lock-breaking, repairing and key-crafting to separate
files. Utilising the new vehicle interaction callback.

Vehicles can be locked using a Locksmith Kit. The kit is used with the
vehicle then a key is returned with an extra-data number that matches
that of the vehicle.

Fixed a small bug that was hiding the action-text UI element when
releasing F/Enter (key 16). It was in the Defences script.