Small admin/database fixes
Fixed an error in the account create SQL statement.

Fixed an error with SetVehicleContainer causing an invalid array access
error if an out of bounds value (such as INVALID_CONTAINER_ID for
vehicles that don't have trunks) is passed.

Fixed a bug that caused players to be locked out of their cars if they
disconnected while in a locked car.

Updated the "/motd" command to properly use DJson instead of y_ini.