Moved lobby/login position, fixed hack detect bug
Moved the lobby/login position from high in the sky to far out in the
ocean. This is the position the player is placed at when they are on the
login/register or create character screen.

Changed the factory spawn to allow players to enter it again (this map
file is actually not in the repo, and it's not mine so I can't
redistribute it. This teleport should be removed. Because of this I will
try and make an improved map management system so teleports can be
placed in the map file itself instead of in the script)

Fixed a small bug with vehicles not having their health reset to 990
after a hacker has healed it. This resulted in vehicles with 1000 health
lying around and banning innocent players who entered.