Small fixes to campfire, defences and safeboxes
Fixed a small bug that would allow the campfire to be picked up instead
of the food on it.

Added button relational ID indexing to defences to eliminate loops from
the lookup code.

Added a HoldActionUpdate hook that stops building if the item is picked
up or destroyed. This previously caused an error and the defence part
would disappear if picked up while being built.