Major update to player data files
IMPORTANT: This update changes the file structure of player character
and player inventory files. When you first run this version, set
"file-check" to 1 in "Settings.json" which will convert all userfiles to
the new format as well as perform some cleanup operations (removing
non-existent accounts, removing files that shouldn't be in that folder
so that, changing the file extension for inventory files from .inv to
.dat)

Moved the bag-type cell to character data instead of inventory data.

Switched the load order to: char>inv instead of inv>char. Because the
bag type must be loaded from the char file before the inventory is
loaded and placed in the bag.

Added 'FileCheck.pwn' script that will run a file check on player data
files after the server starts up. These checks include:
Extension change - inventory files are now .dat instead of silly .inv
Removal if corresponding file is missing - if there is a character file
and no matching inventory file, it will be removed (and vice versa)
Removal if account nonexistent - file is removed if the name on it isn't
in the accounts database.

There is further explanation in the comment at the top of the FileCheck
script.

Added two work-in-progress scripts, ItemStream which will be used to
convert a list of items into a string of cells with varying lengths of
extra-data as well as converting a string back to a set of items and
applying the extradata correctly. This will likely happen after a SIF
update that will change extra-data to a string instead of a single cell.
Secondly, Statistics will record hourly stats as well as continuous
counters of various stats from the server and store them inside a
separate SQLite database for collation and graphing.

Added a "file-check" setting to settings.json which, if true, will
perform the file check and conversion operations on player character and
inventory files on startup. This isn't required all the time so keep it
off otherwise it will slow down restarts and startups (7k accounts took
around 4 minutes on an i7-2600)

Added a small utility script that converts the log files to the new
filename format (year-month-day instead of day-month-year, which would
cause major organisational problems and make finding a certain day
difficult) just run this once on old log files only (move the updated
ones if you ran the server since the log file update. These files should
be easy to distinguish as they would start with a two digit number
instead of '2013')